+++ transformers/storage/median/initializer/initializer.go
@@ -4,13 +4,13 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/makerdao/vdb-mcd-transformers/transformers/shared/constants"
 	mcdStorage "github.com/makerdao/vdb-mcd-transformers/transformers/storage"
-	"github.com/makerdao/vdb-mcd-transformers/transformers/storage/medianizer"
+	"github.com/makerdao/vdb-mcd-transformers/transformers/storage/median"
 	"github.com/makerdao/vulcanizedb/libraries/shared/factories/storage"
 )
 
 var medianizerAddress = constants.GetContractAddress("MCD_MEDIANIZER")
 var StorageTransformerInitializer storage.TransformerInitializer = storage.Transformer{
 	Address:           common.HexToAddress(medianizerAddress),
-	StorageKeysLookup: storage.NewKeysLookup(medianizer.NewKeysLoader(&mcdStorage.MakerStorageRepository{}, medianizerAddress)),
-	Repository:        &medianizer.MedianizerStorageRepository{ContractAddress: medianizerAddress},
+	StorageKeysLookup: storage.NewKeysLookup(median.NewKeysLoader(&mcdStorage.MakerStorageRepository{}, medianizerAddress)),
+	Repository:        &median.MedianizerStorageRepository{ContractAddress: medianizerAddress},
 }.NewTransformer
