+++ transformers/storage/median/keys_loader_test.go
@@ -1,8 +1,8 @@
-package medianizer_test
+package median_test
 
 import (
 	mcdStorage "github.com/makerdao/vdb-mcd-transformers/transformers/storage"
-	"github.com/makerdao/vdb-mcd-transformers/transformers/storage/medianizer"
+	"github.com/makerdao/vdb-mcd-transformers/transformers/storage/median"
 	"github.com/makerdao/vdb-mcd-transformers/transformers/storage/test_helpers"
 	"github.com/makerdao/vdb-mcd-transformers/transformers/test_data"
 	"github.com/makerdao/vulcanizedb/libraries/shared/factories/storage"
@@ -18,20 +18,20 @@ var _ = Describe("Medianizer storage keys loader", func() {
 	)
 	BeforeEach(func() {
 		storageRepository = &test_helpers.MockMakerStorageRepository{}
-		storageKeysLoader = medianizer.NewKeysLoader(storageRepository, test_data.MedianizerAddress())
+		storageKeysLoader = median.NewKeysLoader(storageRepository, test_data.MedianizerAddress())
 	})
 
 	It("returns value metadata for static keys", func() {
 		mappings, err := storageKeysLoader.LoadMappings()
 		Expect(err).NotTo(HaveOccurred())
 
-		Expect(mappings[medianizer.ValAndAgeStorageKey]).To(Equal(types.ValueMetadata{
+		Expect(mappings[median.ValAndAgeStorageKey]).To(Equal(types.ValueMetadata{
 			Name:        mcdStorage.Packed,
 			Type:        types.PackedSlot,
 			PackedTypes: map[int]types.ValueType{0: types.Uint128, 1: types.Uint48},
-			PackedNames: map[int]string{0: medianizer.Val, 1: medianizer.Age},
+			PackedNames: map[int]string{0: median.Val, 1: median.Age},
 		}))
 
-		Expect(mappings[medianizer.BarKey]).To(Equal(medianizer.BarMetadata))
+		Expect(mappings[median.BarKey]).To(Equal(median.BarMetadata))
 	})
 })
