+++ transformers/storage/median/medianizer_suite_test.go
@@ -1,16 +1,17 @@
-package medianizer
+package median
 
 import (
+	"io/ioutil"
+	"testing"
+
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	"github.com/sirupsen/logrus"
-	"io/ioutil"
-	"testing"
 )
 
 func TestMedianizer(t *testing.T) {
 	RegisterFailHandler(Fail)
-	RunSpecs(t, "Medianzier Suite")
+	RunSpecs(t, "Median Suite")
 }
 
 var _ = BeforeSuite(func() {
